added some product crud

+++ Source/Web/App/Models/Product.cs
@@ -20,8 +20,8 @@ namespace MvcMovie.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public float Price { get; set; }
-        public float MemberPrice { get; set; } 
+        public decimal Price { get; set; }
+        public decimal MemberPrice { get; set; } 
         public System.DateTime UpdatedAt { get; set; } 
     }
